requirements.txt
google-generativeai
pandas
sqlalchemy
jsonschema
dotenv 
os